requirements.txt
# Core Streamlit
streamlit>=1.31.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# PDF Parsing
pdfplumber>=0.10.0

# Word Document Parsing
python-docx>=1.1.0

# Visualization
plotly>=5.18.0

# API Requests
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Optional Database
duckdb>=0.9.0

# Date/Time utilities
python-dateutil>=2.8.0

# Data validation
jsonschema>=4.20.0
